Speech_To_Text (Offline Using Openai's Whisper)/Reatime/Method 1/requirement.txt
openai-whisper==20240930
PyAudio==0.2.14
SpeechRecognition==3.11.0

ffmpeg (Version 7.1 used)
